Secretaria 36

See merge request !24

+++ src/main/java/es/tenerife/secretaria/libro/service/impl/RolServiceImpl.java
@@ -1,7 +1,7 @@
 
 package es.tenerife.secretaria.libro.service.impl;
 
-import java.util.HashSet;
+import java.util.List;
 import java.util.Set;
 
 import org.hibernate.criterion.DetachedCriteria;
@@ -9,13 +9,12 @@ import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 import org.springframework.stereotype.Service;
 
-import com.google.common.collect.Lists;
-
 import es.tenerife.secretaria.libro.domain.Rol;
 import es.tenerife.secretaria.libro.repository.RolRepository;
 import es.tenerife.secretaria.libro.repository.UsuarioRepository;
 import es.tenerife.secretaria.libro.service.RolService;
 import es.tenerife.secretaria.libro.web.rest.errors.CustomParameterizedException;
+import es.tenerife.secretaria.libro.web.rest.errors.ErrorConstants;
 import es.tenerife.secretaria.libro.web.rest.util.QueryUtil;
 
 @Service
@@ -42,10 +41,10 @@ public class RolServiceImpl implements RolService {
 	}
 
 	@Override
-	public Set<Rol> findAll(String query) {
+	public List<Rol> findAll(String query) {
 		log.debug("Petición para buscar roles con query {}", query);
 		DetachedCriteria criteria = queryUtil.queryToRolCriteria(query);
-		return new HashSet<>(Lists.newArrayList(rolRepository.findAll(criteria)));
+		return rolRepository.findAll(criteria);
 	}
 
 	@Override
@@ -67,6 +66,9 @@ public class RolServiceImpl implements RolService {
 			throw new CustomParameterizedException("error.palabraClave.users-has-role", codigo);
 		}
 		Rol rol = rolRepository.findOneByCodigo(codigo);
+		if (rol == null) {
+			throw new CustomParameterizedException(ErrorConstants.ENTIDAD_NO_ENCONTRADA, codigo);
+		}
 		rolRepository.delete(rol.getId());
 	}
 
@@ -77,7 +79,7 @@ public class RolServiceImpl implements RolService {
 	}
 
 	@Override
-	public Set<Rol> findByOperacion(Long operacionId) {
+	public List<Rol> findByOperacion(Long operacionId) {
 		log.debug("Petición para buscar roles de usuario {}", operacionId);
 		return rolRepository.findByOperacionesId(operacionId);
 	}
